requirements.txt
absl-py==0.8.1
alabaster==0.7.12
ansiwrap==0.8.4
arrow==0.15.2
asn1crypto==0.24.0
astor==0.7.1
astroid==2.3.1
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcolz==1.2.1
beautifulsoup4==4.8.0
binaryornot==0.4.4
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
blessings==1.7
blis==0.2.4
bokeh==1.3.4
boto==2.49.0
Bottleneck==1.2.1
cachetools==3.1.1
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
conda==4.7.12
conda-package-handling==1.3.11
configparser==4.0.2
confuse==1.0.0
contextlib2==0.6.0
cookiecutter==1.6.0
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.0
dataclasses==0.6
datalab==1.1.5
decorator==4.4.0
defusedxml==0.6.0
distributed==2.5.1
docker==4.1.0
docutils==0.14
entrypoints==0.3
enum34==1.1.6
et-xmlfile==1.0.1
fairing==0.5.3
fast-histogram==0.7
fastai==1.0.58
fastcache==1.1.0
fastprogress==0.1.21
filelock==3.0.12
Flask==1.1.1
fsspec==0.5.1
future==0.17.1
gast==0.2.0
gcsfs==0.3.0
gevent==1.4.0
gitdb2==2.0.6
GitPython==3.0.3
glob2==0.7
gmpy2==2.0.8
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.20.0
google-cloud-core==1.0.3
google-cloud-dataproc==0.5.0
google-cloud-datastore==1.9.0
google-cloud-language==1.3.0
google-cloud-logging==1.13.0
google-cloud-monitoring==0.31.1
google-cloud-spanner==1.10.0
google-cloud-storage==1.20.0
google-cloud-translate==1.6.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
gpustat==0.5.0
greenlet==0.4.15
grpc-google-iam-v1==0.12.3
grpcio==1.25.0
gym==0.10.9
h5py==2.8.0
HeapDict==1.0.1
html5lib==1.0.1
htmlmin==0.1.12
httplib2==0.14.0
idna==2.7
imageio==2.4.1
imagesize==1.1.0
imgaug==0.2.1
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jeepney==0.4.1
Jinja2==2.10.1
jinja2-time==0.2.0
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
julia==0.2.0
jupyter==1.0.0
jupyter-aihub-deploy-extension==0.1
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.5.0
jupyter-highlight-selected-word==0.2.0
jupyter-http-over-ws==0.0.6
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==1.1.4
jupyterlab-git==0.8.1
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.0.1
kubernetes==10.0.1
lazy-object-proxy==1.4.2
libarchive-c==2.8
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
murmurhash==1.0.2
nb-conda==2.2.1
nb-conda-kernels==2.2.2
nbconvert==5.6.0
nbdime==1.1.0
nbformat==4.4.0
nbpresent==3.0.2
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
notebook-executor==0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.17.4
numpydoc==0.9.1
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.1.26
openpyxl==3.0.0
packaging==19.2
pandas==0.25.1
pandas-profiling==2.3.0
pandocfilters==1.4.2
papermill==1.2.0
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
phik==0.9.8
pickleshare==0.7.5
Pillow==5.4.1
Pillow-SIMD==6.0.0.post0
pkginfo==1.5.0.1
plac==0.9.6
plotly==4.1.1
pluggy==0.13.0
ply==3.11
poyo==0.5.0
preshed==2.0.1
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.10.0
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pyarrow==0.14.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pydot==1.4.1
-e git+https://github.com/cdrckrgt/PyFEBOL@2cdb10d7f90fef79e9663dd75c2271792dc47693#egg=PyFEBOL
pyflakes==2.1.1
pyglet==1.3.2
Pygments==2.4.2
pylint==2.4.2
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.3.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
pyrsistent==0.15.4
PySocks==1.7.0
pytest==5.2.0
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-pylint==0.14.1
pytest-remotedata==0.3.2
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
readme-renderer==24.0
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
retrying==1.3.3
rope==0.14.0
rsa==4.0
ruamel-yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.1
scipy==1.1.0
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.13.0
smmap2==2.0.5
snowballstemmer==1.9.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
spacy==2.1.8
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.8
sqlparse==0.3.0
srsly==0.1.0
statsmodels==0.10.1
sympy==1.4
tables==3.5.2
tblib==1.4.0
tenacity==5.1.1
tensorboard==1.14.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
textwrap3==0.9.2
thinc==7.0.8
toolz==0.10.0
torch==1.1.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.32.2
traitlets==4.3.2
twine==1.13.0
typed-ast==1.4.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.2
virtualenv==16.7.5
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwt==1.3.0
zict==1.0.0
zipp==0.6.0
